adds DashboardWidget
Adds a standard DashbaordWidget component and remove the unused Notifications comp. from the render method
